code formatting, better naming
